feat: explicit webcomponent registration path
