+++ packages/malloy-render/package.json
@@ -16,6 +16,7 @@
     ".": "./dist/module/index.umd.js",
     "./style": "./dist/module/style.css",
     "./webcomponent": "./dist/webcomponent/malloy-render.umd.js",
+    "./webcomponent/register": "./dist/register/register.umd.js",
     "./webcomponent/style": "./dist/webcomponent/style.css"
   },
   "scripts": {
@@ -29,7 +30,7 @@
     "storybook": "storybook dev -p 6006",
     "build-storybook": "storybook build",
     "build-source": "vite build --outDir 'dist/module' --config vite.config.ts",
-    "build-webcomponent": "vite build --outDir 'dist/webcomponent' --config vite.config.webcomponent.ts",
+    "build-webcomponent": "vite build --outDir 'dist/webcomponent' --config vite.config.webcomponent.ts && vite build --outDir 'dist/register' --config vite.config.webcomponent-register.ts",
     "build-types": "tsc --build --declaration --emitDeclarationOnly"
   },
   "dependencies": {
